Move Report Issue form to tab nav and update style
