+++ src/assets/svgs/TabBarNav/questionMark.ts
@@ -1,7 +1,8 @@
 export default `
 <svg width="25" height="25" viewBox="0 0 25 25" xmlns="http://www.w3.org/2000/svg">
-<path d="M12.5 2C14.5767 2 16.6068 2.61581 18.3335 3.76957C20.0602 4.92332 21.406 6.5632 22.2007 8.48182C22.9955 10.4004 23.2034 12.5116 22.7983 14.5484C22.3931 16.5852 21.3931 18.4562 19.9246 19.9246C18.4562 21.3931 16.5853 22.3931 14.5485 22.7982C12.5117 23.2034 10.4005 22.9955 8.48183 22.2007C6.56321 21.406 4.92333 20.0602 3.76957 18.3335C2.61582 16.6068 2.00001 14.5767 2.00001 12.5C2.00001 9.71523 3.10625 7.04451 5.07538 5.07538C7.04452 3.10625 9.71523 2 12.5 2ZM12.5 0C10.0277 0 7.61099 0.733112 5.55538 2.10663C3.49976 3.48015 1.89761 5.43238 0.951511 7.71646C0.00541604 10.0005 -0.242126 12.5139 0.24019 14.9386C0.722505 17.3634 1.91301 19.5907 3.66117 21.3388C5.40933 23.087 7.63661 24.2775 10.0614 24.7598C12.4861 25.2421 14.9995 24.9946 17.2835 24.0485C19.5676 23.1024 21.5199 21.5002 22.8934 19.4446C24.2669 17.389 25 14.9723 25 12.5C25 9.18479 23.683 6.00537 21.3388 3.66117C18.9946 1.31696 15.8152 0 125 0V0Z" />
-<path d="M12.6148 5.875L12.8848 5.9C13.5534 5.94217 14.2076 6.11189 14.8123 6.4C15.3525 6.64866 15.8129 7.04246 16.1423 7.5375C16.3933 7.9274 16.5483 8.37117 16.5948 8.8325C16.6471 9.23143 16.6147 9.6369 16.4998 10.0225C16.3572 10.4826 16.1089 10.9029 15.7748 11.25C15.3927 11.6325 14.9741 11.9766 14.5248 12.2775C14.2748 12.45 14.0523 12.63 13.8173 12.81C13.6047 12.9857 13.4447 13.2167 13.3548 13.4775C13.3048 13.6125 13.2473 13.7475 13.1948 13.8825C13.1304 14.0621 13.0125 14.2175 12.8569 14.328C12.7013 14.4384 12.5156 14.4984 12.3248 14.5C12.1089 14.5237 11.8905 14.4892 11.6923 14.4C11.5424 14.3267 11.4124 14.2181 11.3135 14.0836C11.2147 13.949 11.1499 13.7925 11.1248 13.6275C11.0854 13.3978 11.0916 13.1625 11.1431 12.9352C11.1945 12.7078 11.2903 12.4928 11.4248 12.3025C11.6412 11.9862 11.9061 11.7061 12.2098 11.4725C12.5373 11.205 12.8748 10.95 13.2098 10.6825C13.4879 10.4697 13.7207 10.2037 13.8948 9.89999C14.0301 9.64231 14.0857 9.35023 14.0545 9.06088C14.0233 8.77153 13.9069 8.49798 13.7198 8.27499C13.5216 8.05886 13.2583 7.91318 12.9698 7.86C12.5858 7.77035 12.1856 7.77722 11.8048 7.88C11.6096 7.93332 11.428 8.02785 11.2723 8.15722C11.1167 8.28659 10.9905 8.44779 10.9023 8.63C10.7848 8.85 10.6823 9.07499 10.5723 9.29749C10.4917 9.46575 10.3826 9.61883 10.2498 9.75C10.1348 9.85106 10.0002 9.9274 9.8544 9.97428C9.70861 10.0212 9.55475 10.0376 9.40235 10.0225C9.21499 10.0176 9.03153 9.96789 8.86735 9.87749C8.71617 9.79929 8.5895 9.6809 8.50128 9.53534C8.41307 9.38978 8.36672 9.2227 8.36735 9.0525C8.35456 8.50038 8.53099 7.9605 8.86735 7.5225C9.25622 6.99684 9.7724 6.57873 10.3673 6.3075C10.8161 6.09451 11.2966 5.95603 11.7898 5.8975L12.0698 5.87L12.6148 5.875Z" />
-<path d="M12.1177 19.125C12.0427 19.11 11.9677 19.0975 11.8952 19.0775C11.5651 18.9911 11.2786 18.786 11.0904 18.5014C10.9023 18.2167 10.8258 17.8727 10.8757 17.5352C10.9256 17.1977 11.0983 16.8904 11.3607 16.6724C11.6231 16.4544 11.9568 16.3408 12.2977 16.3536C12.6387 16.3664 12.9629 16.5046 13.2083 16.7417C13.4536 16.9788 13.6028 17.2981 13.6273 17.6384C13.6517 17.9787 13.5497 18.316 13.3408 18.5857C13.1319 18.8555 12.8308 19.0386 12.4952 19.1L12.3727 19.125H12.1177Z" />
+<path d="M12.5 2C14.5767 2 16.6068 2.61581 18.3335 3.76957C20.0602 4.92332 21.406 6.5632 22.2007 8.48182C22.9955 10.4004 23.2034 12.5116 22.7983 14.5484C22.3931 16.5852 21.3931 18.4562 19.9246 19.9246C18.4562 21.3931 16.5853 22.3931 14.5485 22.7982C12.5117 23.2034 10.4005 22.9955 8.48183 22.2007C6.56321 21.406 4.92333 20.0602 3.76957 18.3335C2.61582 16.6068 2.00001 14.5767 2.00001 12.5C2.00001 9.71523 3.10625 7.04451 5.07538 5.07538C7.04452 3.10625 9.71523 2 12.5 2ZM12.5 0C10.0277 0 7.61099 0.733112 5.55538 2.10663C3.49976 3.48015 1.89761 5.43238 0.951511 7.71646C0.00541604 10.0005 -0.242126 12.5139 0.24019 14.9386C0.722505 17.3634 1.91301 19.5907 3.66117 21.3388C5.40933 23.087 7.63661 24.2775 10.0614 24.7598C12.4861 25.2421 14.9995 24.9946 17.2835 24.0485C19.5676 23.1024 21.5199 21.5002 22.8934 19.4446C24.2669 17.389 25 14.9723 25 12.5C25 9.18479 23.683 6.00537 21.3388 3.66117C18.9946 1.31696 15.8152 0 12.5 0Z" />
+<path d="M12.615 5.875L12.885 5.9C13.5535 5.94217 14.2078 6.11189 14.8125 6.4C15.3526 6.64866 15.8131 7.04246 16.1425 7.5375C16.3934 7.9274 16.5485 8.37117 16.595 8.8325C16.6472 9.23143 16.6148 9.6369 16.5 10.0225C16.3573 10.4826 16.109 10.9029 15.775 11.25C15.3929 11.6325 14.9742 11.9766 14.525 12.2775C14.275 12.45 14.0525 12.63 13.8175 12.81C13.6049 12.9857 13.4448 13.2167 13.355 13.4775C13.305 13.6125 13.2475 13.7475 13.195 13.8825C13.1305 14.0621 13.0126 14.2175 12.857 14.328C12.7014 14.4384 12.5157 14.4984 12.325 14.5C12.109 14.5237 11.8906 14.4892 11.6925 14.4C11.5425 14.3267 11.4125 14.2181 11.3136 14.0836C11.2148 13.949 11.1501 13.7925 11.125 13.6275C11.0855 13.3978 11.0917 13.1625 11.1432 12.9352C11.1947 12.7078 11.2904 12.4928 11.425 12.3025C11.6413 11.9862 11.9062 11.7061 12.21 11.4725C12.5375 11.205 12.875 10.95 13.21 10.6825C13.488 10.4697 13.7208 10.2037 13.895 9.89999C14.0302 9.64231 14.0858 9.35023 14.0546 9.06088C14.0235 8.77153 13.907 8.49798 13.72 8.27499C13.5217 8.05886 13.2584 7.91318 12.97 7.86C12.5859 7.77035 12.1857 7.77722 11.805 7.88C11.6097 7.93332 11.4281 8.02785 11.2725 8.15722C11.1168 8.28659 10.9906 8.44779 10.9025 8.63C10.785 8.85 10.6825 9.07499 10.5725 9.29749C10.4918 9.46575 10.3827 9.61883 10.25 9.75C10.1349 9.85106 10.0003 9.9274 9.85453 9.97428C9.70874 10.0212 9.55487 10.0376 9.40247 10.0225C9.21511 10.0176 9.03165 9.96789 8.86747 9.87749C8.71629 9.79929 8.58962 9.6809 8.50141 9.53534C8.41319 9.38978 8.36684 9.2227 8.36747 9.0525C8.35468 8.50038 8.53111 7.9605 8.86747 7.5225C9.25634 6.99684 9.77252 6.57873 10.3675 6.3075C10.8162 6.09451 11.2967 5.95603 11.79 5.8975L12.07 5.87L12.615 5.875Z" />
+<path d="M12.1176 19.125C12.0426 19.11 11.9676 19.0975 11.8951 19.0775C11.565 18.9911 11.2784 18.786 11.0903 18.5014C10.9022 18.2167 10.8257 17.8727 10.8756 17.5352C10.9255 17.1977 11.0982 16.8904 11.3606 16.6724C11.623 16.4544 11.9567 16.3408 12.2976 16.3536C12.6386 16.3664 12.9628 16.5046 13.2081 16.7417C13.4535 16.9788 13.6027 17.2981 13.6272 17.6384C13.6516 17.9787 13.5496 18.316 13.3407 18.5857C13.1318 18.8555 12.8307 19.0386 12.4951 19.1L12.3726 19.125H12.1176Z" />
 </svg>
+
 `
