+++ src/assets/svgs/alertCircle.ts
@@ -1,0 +1,6 @@
+export default `
+<svg width="22" height="22" viewBox="0 0 22 22" xmlns="http://www.w3.org/2000/svg">
+<path fill-rule="evenodd" clip-rule="evenodd" d="M11 2C6.02944 2 2 6.02944 2 11C2 15.9706 6.02944 20 11 20C15.9706 20 20 15.9706 20 11C20 6.02944 15.9706 2 11 2ZM0 11C0 4.92487 4.92487 0 11 0C17.0751 0 22 4.92487 22 11C22 17.0751 17.0751 22 11 22C4.92487 22 0 17.0751 0 11ZM11 6C11.5523 6 12 6.44772 12 7V11C12 11.5523 11.5523 12 11 12C10.4477 12 10 11.5523 10 11V7C10 6.44772 10.4477 6 11 6ZM10 15C10 14.4477 10.4477 14 11 14H11.01C11.5623 14 12.01 14.4477 12.01 15C12.01 15.5523 11.5623 16 11.01 16H11C10.4477 16 10 15.5523 10 15Z" />
+</svg>
+
+`
