+++ src/styles/headers.ts
@@ -4,10 +4,11 @@ import * as Colors from "./colors"
 import * as Typography from "./typography"
 
 export const headerStyle: ViewStyle = {
-  backgroundColor: Colors.headerBackground,
+  backgroundColor: Colors.tertiaryBlue,
 }
 
 export const headerTitleStyle: TextStyle = {
+  ...Typography.base,
   color: Colors.headerText,
   letterSpacing: Typography.mediumLetterSpacing,
   textTransform: "uppercase",
